+++ include/ZipDeflate.h
@@ -6,6 +6,7 @@
 #include<cstdint>
 #include<vector>
 #include<memory>
+#include<cstddef>
 
 // Class xử lý giải nén dòng bit Deflate
 class ZipDeflate {
@@ -13,10 +14,10 @@ public:
     ZipDeflate(std::vector<uint8_t> const& data) {
         bitStream = std::unique_ptr<ZipBitStreamInterface>(new ZipBitStream(data));
     }
-    ZipDeflate(std::vector<uint8_t> const& data, int _start_offset, int _end_offset) {
+    ZipDeflate(std::vector<uint8_t> const& data, size_t _start_offset, size_t _end_offset) {
         bitStream = std::unique_ptr<ZipBitStreamInterface>(new ZipBitStream(data, _start_offset, _end_offset));
     }
-    ZipDeflate(std::vector<uint8_t> const& data, int _start_offset, int _end_offset, ZipPassword initKey) {
+    ZipDeflate(std::vector<uint8_t> const& data, size_t _start_offset, size_t _end_offset, ZipPassword initKey) {
         bitStream = std::unique_ptr<ZipBitStreamInterface>(new ZipBitStreamEncrypted(data, _start_offset, _end_offset, initKey));
     }
     std::vector<uint8_t> Decode();
